feat: resumable uploads
Store ETags for uploaded chunks in a localStorage. Check if there is a ETag in cache before uploading a chunk.
Clean cache on cancel or complete. Don't clean the cache on error (so we can actually resume later).

This implementation is not perfect and requires that client reuses same uploadId (caching it on the client).